Add PSI bridge for Java files, and other lint fixes
Adds a new bridge from IntelliJ's PSI to Lombok AST. This lets us drop
the much slower default LombokParser and indirectly use the PSI data
for lint checks instead, which is a lot faster.

Because of that, this check also adds Java checks back into the
background annotator. It also fixes the routine which determines
whether a lint check requires batch mode to consider lint's analysis
scopes, and it tweaks the error output to use HTML markup (for keyword
highlighting, URLs etc).

It also tweaks the R file header comment.

Change-Id: I04f45d5f318cb5459daec34c5ff49b5b628fb5c4