Fix px versus dp handling
Further zoom fixes in the Android designer.

First, convert between screen and model coordinates such that when for
example resizing, the size being set is in model coordinates, not
(possibly scaled) screen coordinates. Similarly, when operating on
baselines and margin bounds, which are defined in model coordinates,
translate to view coordinates before mixing with other view bounds.

Second, add utility methods to convert from px to dp (which involves
scaling by the device dpi), and use that when resizing and adjusting
bounds (since the computed bounds are in pixels, and the XML is
setting dp sizes).

Change-Id: I709346a0295e37a40a2fca2b34089b643ef67882