Add drawing styles abstraction
This adds a DrawingStyle enumerated type which captures the
different types of visual feedback used by the layout editor,
and then individual feedback and decorator code refers to
these styles rather than specific colors and strokes when
painting.

This also changes the visual style to the ADT layout editor
style.

Change-Id: I1c7d00f778af6d6ffb4773121820fc6779af4b59