logcat: show package name and thread id
This CL maps the pid of the application generating a log to its
package name and annotates the log message with the package name.
The thread id is added as well.

The code that formats the message and the code that decodes the
formatted message later have both been moved into a single class.

Change-Id: Icac71d65cd38810b51bfe98cc22ecfdd1242cc46