Avoid linear layout gravity resizing with small bounds
If you drag over a linear layout, the layout editor will also
attempt to pick a gravity. In a vertical layout, if you drag
near the right side of the layout it will offer to insert the
child with gravity="right" and so on.

However, when you are dropping into a tiny LinearLayout, this
is confusing; just move a pixel or two and the gravity changes
(and it's hard to guess what it will do).

This changeset tweaks this such that if the LinearLayout is
narrower in the given gravity dimension than a certain threshold,
then it will not set the gravity and use the default instead.

Change-Id: Iee59c3696cdfaeb8f93baebc4e54e5bda789d4ef