Add AndroidTestCase#requireRecentSdk
By overriding requireRecentSdk() a unit test can state that it
requires a recent SDK to be loaded, rather than the currently bundled
1.5 SDK (which among other things is missing layoutlib).

For now, setting this flag will force the test to attempt loading the
platform from $ADT_TEST_SDK_PATH and (layoutlib from
$ADT_TEST_SDK_PATH/platforms/$ADT_TEST_PLATFORM), which must be
defined as environment variables or system properties.

Later on we can consider updating the bundled test SDK to something
more recent, but it's currently a stripped down version (5MB) versus a
"real" install size for layoutlib with full resources and fonts nearly
90MB, and that's the platform dir alone, not including other data like
the api versions over in platform-tools.

Change-Id: I80dbb2ca4211394d2bcbd3ba3aac4ffb5e39289c