Config Toolbar: Finish Activity menu and Orientation menu
This changeset adds actions to the activity menu (for assigning an
associated activity for the current layout, by looking up which activities
in the project reference the given R.layout field, as well as an arbitrary
action to pick any Activity), as well as an action for jumping to the Java
class for the currently selected activity.

It also hooks up the UI Mode and Night Mode context menus in the Orientation
menu.

Finally, it rewrites the ManifestInfo class to use PSI directly rather than
relying on File IO + XML parsing to look up manifest data such as which
themes are registered for which activities, and fixes version codename
handling. Plus tests.

Change-Id: Ie37cd3a8cb7c39fa44a09db92b6ec41a60203306