Make Zoom to Fit layout editor operation a mode
This changeset makes the Zoom to Fit operation in the layout editor
behave as a mode, just like the XML Layout Preview already did, such
that these two are consistent. The layout editor zoom support already
did this in a few implicit cases (e.g. when you switched device
orientation) so this just makes the behavior a bit more consistent
(e.g. when maximizing the editor).

Also fix a few bugs related to shadow painting: When zooming in, must
draw drop shadow manually (since we don't pre-create a zoomed image),
and also offset the scaled image a bit when squeezed near the borders
(no padding) in order to make the image look more balanced when there
is a drop shadow.

Change-Id: I4d77c8b3a5cec202060ab9c0d8bf5133610951f5