Make the Android color annotator work for resource references and layouts
First, this makes the Android color annotator work for resource references
such as @color/button as well as framework resources, such as
@android:color/holo_blue_dark.

Second, it makes the annotator work for XML attributes as well (such as
a background= attribute in a layout file) in addition to the existing
support for text values (in res/color files).

Third, it removes support for color constants; as far as I can tell the
spec doesn't allow that. It also moves the color parsing code over to
ResourceHelper (and replaces the existing parseColor method which was
there for the resource chooser.)

Change-Id: If280cd427069cbdde75da550f97250aa153b1142