Replace "Project name" with "Module name"
In the New Project and New Module wizards, rename "Project name" in
the field and error messages to "Module name". This is clearly the
correct thing for New Module. For New Project, it's not incorrect,
as the name is used for both the project and module names. In NPW,
"Project location" is unchanged, and this field doesn't appear at
all in NMW.

Change-Id: I815425675ea80eddf0c991e113971af993e9f524