Support custom views in gradle projects
This CL adds basic support for custom views in Gradle projects.
(Custom views from libraries was already working, as was custom views
in legacy projects).

This is slightly hacky at the moment because it doesn't have access to
the output folder for the gradle project, but it makes a guess based
on the location of the output .apk file, which it does know about.

Change-Id: I754037fdc6399437ac339490b8cfc831a269bddf