PSI-based project resources
This changeset replaces the ProjectResource implementation, currently
based on File processing using the res2 ResourceRepository, with a
custom implementations built directly on top of the Intellij PSI
datastructures.

This allows the repository to be updated incrementally. For example,
editing text within a <string> element in a value file will leave the
resource repository alone and will just flush the ResoureValue for the
corresponding ResourceItem. Editing the name attribute of an item will
remove the old item and create a new one. etc. etc.

Using PSI directly also allows the resource repository to return the
current results even before files are saved, and it also allowed me to
remove all the sync() logic dedicated to seeing if any files had been
marked as stale.

There is a new ResourceFolderRepository which keeps track of the
resources in a given resource folder. The ProjectResources are
computed as a combination of these (one from each flavor and build
type). This means that for example switching flavors is quick; just
reordering these folder repositories (until now, it caused a global
flush). In a followup CL this will also allow us to query *all*
project resources across flavors, not just the currently available
resources in the current flavor.

There is some work left to do; I still want to make the
ModuleSetRepository (which represents a ModuleRepository along with
all its library dependencies) handle module dependency changes.
I'd also like to add some snapshot facility to allow the repository
save and restore itself quickly so we don't have to initialize PSI
trees for all resource files at startup.

Change-Id: I878986d834e8aa284ee0b44616d6974a0e0bdd7b