ScrollView support: Show full client view, not viewport
If you open a layout that has a ScrollView as the root layout,
use a special rendering mode which measures the required height,
and uses that instead of the device height. This allows you to
design layouts that scroll, since otherwise it's tricky to view
or edit anything beyond the portion near the top that is visible.
56804: A Scrollable layout can't be scrolled inside preview

Change-Id: Ib1925a13a62deef5a1353dde301be8374f63d915