Allow testing library modules in run configuration
The run configurations show a combo box with the list of modules
that can actually be run. The list of modules shown there was
coded to never show library modules.

This CL modifies the condition so that library modules are
displayed for test configurations, if the modules are Gradle based.

Change-Id: Icebc223bc90abf5699d746861831a2b07f386f39