Creates Studio modules from folders that do not map to Gradle project.
We stopped creating modules from this type of folders because of the bug where
Gradle complained that task "assemble" was not declared for a module. The
problem with the previous fix is that users will no longer be able to see
files and folders that are inside the folders that were previously used as
module roots.

This CL partially reverts that, but it does not add the "Gradle" facet to
the modules that do not map to Gradle projects. This way JPS will not try to
build these modules with Gradle.

Change-Id: Ifc3bc9448b03c60958c09635e9501f75bf14409b