Converts New Module Wizard to ModuleBuilder infrastructure
Makes the New Module Wizard a ModuleBuilder-based implementation.
It can now be invoked via the add module button in the Project
Structure dialog. A ProjectTemplateFactory class is added so that
projects that have an Android-Gradle facet will use this builder
for adding new modules. The previous AndroidModuleBuilder's
ProjectTemplateFactory is modified so that it is used only in
projects not having an Android-Gradle facet. This should do the
right thing in both environments.

Change-Id: Id7e2077a96544091dcf35b94bc90295bf2193de7