Support project-wide locale and render target settings
The locale setting in the layout preview and layout editors now apply to the
whole project (except for layouts located in locale specific folders, such as
layout-en.)

Ditto for render targets. In addition, there is a new "Automatically Pick Best"
menu item which causes the best target to be picked when the layout is opened
(which means that even if you've been rendering with render target 17, if you
install 18, we'll automatically switch to that rather than keeping the layout
stuck at 17.)

Also work around looper timing issue in the render library.

Change-Id: If992df0352d51ba39674d7ce9497c14c1b5fd6f2