RTL refactoring tweaks
Remove the AndroidRefactoringManager and just invoke the
RtlSupportManager directly.

Fix the labels returned by the UsageViewDescriptor such that
the label there doesn't show getCodeReferencesText() as the
header.

Reformat the code such that it complies with the plugin code
style (2 space indent, else keyword on new line, wildcard imports
when >= 5 items in same package/class.)

Change-Id: Ife7eae7d7595c2afe053a94b9351171895e4edca