Open layout preview rather than  editor, and select in project view
This changeset tweaks the open hook in the template machinery such
that for XML files, it opens the XML text editor rather than the
default editor.  This has the net effect of showing the layout XML
preview instead of the editor, since the editor isn't quite done yet.

It also selects the topmost file in the project view, to make it more
obvious how to navigate the project (since this will locate the main
res folder inside the default module and expose its contents.)

Change-Id: Iefce7959544cfa1b580b6bb7bfda77301fe1df7d