Fix "Namespace '' not bound" XML warnings
The IntelliJ XML editor support would always flag the root element in
XML files with a "Namespace '' not bound" warning. For file types like
XHTML it can find a namespace declaration right on the root element, but
this is typically not included in Android, so work around this by
providing a dummy namespace prefix in these cases.

Change-Id: I0504a66f013ebf2063c6d2594994c25c2d24f4ca