In render errors, hyperlink to sources and include positions
When an exception is thrown during render errors, we show the
stack frame. For exceptions in user code, we added hyperlinks
to point to the source, but for other frames the file and line
number was not included, which makes it harder to track down
the root cause; this CL fixes that.

It also looks to see if the user has the platform source code
installed for the current render target, and if so, it links
directly to it. The hyperlink resolver is also enhanced to handle
line numbers and colum numbers tacked to the end of the URL.

Change-Id: I62bcb9cdc74a29ba0bcb429168c13871fdc70596