traceview: Enable sorting by column
This CL enables sorting of the traceview stats. Since the stats are
displayed in a TreeTable, sorting across different levels of a tree
is ill defined. This particular implementation, passes on the sort
request to the model, and currently, the model interprets this as a
request to sort the sequence of methods displaying profile data.

Change-Id: I3b2c9e6f7f27630bf445dc39b79326e3ce6e5f31