Show progress indicators in the layout preview panel
This changeset adds a progress indicator to the layout preview panel.
(It used to have one as part of the title bar, which was removed when
we added device frames.)

The progress indicator now takes two forms: A large progress indicator
in the center of the rendering panel when it is empty, and a small
progress indicator icon in the upper right hand corner when there is
already a layout rendering present. This is necessary since for
example under Darcula, the progress icon is white, which makes it
indivible on top of a light theme rendering in the preview panel.

Furtheremore, there is an extra progress indicator now which tracks
editor changes, and fires immediately on file switch. This means that
when you're editing within a file, you don't get a progress indicator
between each edit and the render shortly thereafter, but when you're
switching files you can instantly see that the contents are stale.

Some misc code cleanup.

Change-Id: Idf57421fe0899164e55700968419100adf80a36a