logcat: Filter based on currently selected client.
This CL sets the logcat filter to only show messages from
the currently selected VM (by pid) when a VM is selected in the
device panel.

When this occurs, the filter name displayed on the combo box is
set to empty. This allows us to not show this transient filter.
This filter is also not persisted.

This filtering based on selected process can be enabled/disabled
via a toggle action that is displayed on the logcat toolbar.

Fixes https://code.google.com/p/android/issues/detail?id=58534

Change-Id: I234d852d1e846ccc9b91b7d3fd4f5fbe60780e7e