Layout XML pull parser cleanup
First, in the layoutlib parser factory, check whether the file we need
a parser for is in the project, and if so, attempt to create a
PSI-based pull parser for it.  This means that if your layout includes
another layout in your project, and you've edited that included
layout, we'll render the current contents of the included layout rather
than what is saved on disk.

Second, our layout pull parser had support for automatically increasing
the padding of a given set of nodes. Most of the time it is not used.
Therefore, I pulled all the padding related logic into a separate
subclass which is only used when padding is enabled. This simplifies
the logic of the PSI pull parser to really only deal with PSI pull
parsing.

I also took the opportunity to rename the class, since it's only
suitable for layout parsing, so the much more general name
XmlTagPullParser was a bit misleading. When we support rendering of
more than layouts, we should extract out the generic parsing bits into
an XmlTagPullParser class, and leave only the layout specific bits in
LayoutPsiPullParser. Similarly, the ContextPullParser is now called
LayoutFilePullParser, since it's like LayoutPsiPullParser but operates
on files rather than PSI elements.

Change-Id: If27355918f3bd20744ebd48c21184c20ce1aa1ae