Added HTTP proxy settings to Gradle daemon used for building.
HTTP proxy settings are already being set for importing
projects.

It is necessary to keep system properties between the process
that imports a project and the one that builds a project. This
way they can use only one Gradle daemon. If system properties
differ, Gradle will start a new daemon, which uses a
significant amount of energy.

Change-Id: I33894575c306c028cc4c4ab9edaf61519bf627a7