Activate correct device in DDMS on app launch
Launch configurations can optionally reveal the DDMS/logcat
panel. However, the device that is currently selected there
has no relation to the device on which the application was
launched.

This CL just sets the device selected in the DDMS view to be
the same device on which the application was launched.

Fixes: https://code.google.com/p/android/issues/detail?id=60266

Change-Id: I3636d9e4170a5169f0ba3572147d47481d93d4c4