Fix setting Android ToolWindow shortcut.
Android Studio removes the shortcut for the show TODO toolwindow
action, and reuses it to show the Android Tool window.

It turns out that when adding a shortcut for an action, IJ also
copies the shortcut from the parent keymap, if the parent keymap
defines a shortcut.

This CL changes the order in which the keymaps are changed to change
the default keymap last, assuming that all other keymaps inherit
from this default keymap.

Fixes: https://code.google.com/p/android/issues/detail?id=61050

Change-Id: Ifd6ae1873bffcdfecbb9ab153e1566cfaa93e7c8