Custom renderer for ArrayMap objects
This CL adds a custom renderer that is used by the debugger
when displaying ArrayMap objects.

ArrayMap implements the Map interface, but its entry set does
not support the toArray() method. As a result, the default Map
renderer just shows an UnsupportedOperationException while
examining the contents of an ArrayMap from within the debugger.

Change-Id: I44f424672d0ce29c72f65f2852f6de21d6861cf1