ActivityWizard: Fetch minSdkVersion and package from build.gradle
This changeset fetches the minSdkVersion and package from the
Gradle model, and uses these instead of the values from the manifest
file, if provided.

Also tweaks initial New Activity Wizard window size to ensure the
error label isn't cropped, and in merge conflict output places the
old contents in the first section.

Change-Id: Ia185a425c497fb597aee103e90705b15d60106e8