Template unit tests
This CL adds unit tests for template instantiation. It tests both
creating projects along with all the default activities, as well as
creating all the templates (activities and templates like services and
custom views) into existing projects.

For each template, it tests creating the template in a variety of
environments (minSdk, targetSdk and buildApi combinations), along with
various combinations of boolean and enum value parameters. By default,
the unit test just checks a small subset of the possible combinations;
the system property or environment variable
"com.android.tools.idea.templates.TemplateTest.COMPREHENSIVE" can be
set to "true" on the build server to get the test to check more
possibilities.

After each template has been created, it also performs a gradle build
and checks that the build is successful.

A number of templates do not pass these tests; these have been
deliberately added to an exclude list in the test; we should work our
way through the templates and fix them.

This changeset also provides the beginnings of AndroidGradleTestCase;
a superclass for unit tests that want to perform unit tests against a
Gradle project rather than an IntelliJ test, as is the case for the
AndroidTestCase class. This is a work in progress; I need to add the
ability to load existing projects rather than just creating new ones,
and unit tests will need to sync the gradle model which will require
some asynchronous handling.

Change-Id: I8f9cc3d9862fb5e9921ca0d4ae391c9f459cf258