Show Gradle output in console.
This CL adds a "Gradle Console" tool window that
displays Gradle output when:
1. Generating sources
2. "Making" a project
3. "Rebuilding" a project
4. "Compiling" a project

This console only works with the new approach
for invoking Gradle to build a project that does
not use JPS (introduced with CL
https://android-review.googlesource.com/68149 .)

The console shows different text colors for
regular output and errors. The colors are handled
by IDEA and are theme-aware.

This console is *not* automatically activated on
every project build. To activate the console use
the new "Show Console Output" action in the
"Messages" tool window.

Change-Id: I2f447b458e26e5d91d54a4935fcfc607587767a9