61300: Rebuild fails since last update
This CL fixes a bug where opening a layout and then attempting to rebuild
the project does not work on Windows, because
(1) Opening a layout will initialize a URLClassLoader which points to
    all .jar files in AAR dependencies, such that custom views like the
    DrawerLayout in app compat can be found during layout rendering.
(2) On Windows, the classes.jar file referenced by a URLClassLoader will
    be opened in such a way that the file is locked
(3) On Windows, you cannot delete a file if it is in use by a different
    process

This means that when you try to clean/rebuild the project, it cannot
delete the build folder since it contains the exploded AAR contents.

The workaround is to use IntelliJ's local implementation of a class loader
which does not have the locking issue.

Change-Id: Id6b6e35323b71659cac18f9092fe3c0eb7c4cfa3