Module properties editor
Implements a GUI editor for many parts of the build.gradle file,
including general config, build types, flavors, signing configs,
etc. To do this, the BuildFileKey mechanism is extended to add
type information, as needed to make many of the properties work
properly when getting/setting. A new mechanism is added to allow
the code to generically deal with named closures that have
key/value subproperties; many entities work this way. There is UI
to allow editing of these objects in a general way through a table
component. The test suite has been brought up to snuff to provide
coverage of much of the framework code.

Still left to do (will file bugs on these):

* You can't edit the defaultConfig yet. Will be treated ad a flavor.
* There's a type which is a reference to another type; in Groovy,
  this is a named entity like a signingConfig, and when referred to
  elsewhere in the file you do so with a Groovy reference, not a
  string. The code has the idea of a reference type, but it's
  treating it like a dumb string now. It needs to do validation to
  prevent users from entering invalid reference types.
* All NamedObject types are edited in a giant table right now where
  every possible property is a table column. This isn't very friendly
  for complex objects, and I need to implement a scheme where you
  have a separate editor; perhaps you show only the most important
  properties in the editor. I will use the SingleObjectTable/Model
  to do this.
* There needs to be an editor for repository statements (i.e. Maven
  coordinates for the Android-Gradle plugin and other library
  dependencies). I started on one but it fell away.
* There may be other top-level properties in the Android config that
  the user should be able to edit in the general properties panel.
* It isn't good yet about identifying user-generated changes in the
  build.gradle file and either refusing to work with it or gracefully
  preserving those changes. This is already documented in a bug.

Change-Id: I744c12cbf627faac397d9b34ca9534f20c396450