Set classpath and test classpath for Java modules
This CL sets the classpath to "build/classes/main" and the test
classpath to "build/classes/test" when a Java module is imported.
This allows the regular Junit test runner to obtain the appropriate
classpath.

While this CL sets the classpath and allows JUnit tests to be run,
there is still an issue that the tests aren't compiled by the IDE
at all. The fix for that will be in a subsequent CL.

This is a partial fix for https://code.google.com/p/android/issues/detail?id=60916

Change-Id: Iebf3b81a7aa338f4d0f9ff9d1714915c2d946a45