Add size and resize metadata for the layout editor
The resize policy metadata indicates whether a widget is resizable in
the horizontal dimension, the vertical dimension, whether it expects
its aspect ratio to be preserved, etc. This metadata will be used in a
follow-up CL to control which resize control points are shown for
selected widgets, and how they behave during resizing.

It also adds metadata about how widgets expect to be sized initially,
such that for example a horizontal linear layout will always fill its
parent's width, but will only fill vertically if it is in a vertical
parent.

Finally it cleans up imports and moves the Insets class to a new
package.

Change-Id: I3acdcd43859477943279168df549dd986c4734f2