Renames Module "Lib" or "App"
The old module name was the same as the project name for a new
project leading to some confusion. This change defaults to naming
the Module "Lib" or "App" (again with automatic naming conflict
resolution) to help clarify the function of the module.

Also includes a fix for issue 62146 (NewModuleWizard infinite looping)
Change-Id: I8f95807a3600b197ef84067f58b3d9f58021fdf7