Fix for nasty Heisenbug
The PSI hierarchy seems to have non-deterministic behavior around annotations. The (empty)
argument  list of a lone "@Override" annotation sometimes results in the creation of
an AnnotationParamListElement, other times a PsiAnnotationParamListImpl  -
in exactly the the same place. The was hard to track down as only the latter type
seems to appear when running in the debugger.

The unifier is not sensitive to the detailed structure of the PSI elements as it is only
matching one PSI parse against another. But it is sensitive to differences
between two 'parses' of the same input.

This problem seems to be restricted to the parsing of annotation parameters -
which are not needed for the Navigation Editor - so this CL arranges that
annotations are ignored by the unification algorithm outright.

Also fixed (unrelated) threading issue in the code that calls the rendering library.

Change-Id: Iee61a7f6df8ea8d20326267ddc5a1c63d913cbf0