Repository editor
Editing of repository statements for the Android plugin and
separately for user libraries. Allows mavenCentral, mavenLocal,
and URL dependencies. For now, you just type in a comma-delimited
list into a text field; in the future this can be made fancier
with a combobox to make mavenCentral and mavenLocal more
discoverable.

Change-Id: I091312fe9626ab6b391fa971803a8fd001b30e77