Support RTL live layout preview
Support previewing RTL rendering simultaneously (when enabled,
you see the layout and its RTL version next to it).

Also fix the layout editor's RelativeLayout handling such that it
properly sets start/end attributes in addition to left/right (or
instead of, depending on the minSdkVersion) to properly design
layouts for RTL locales.

Change-Id: I0c3de91442e8b0dbcce1b7ec0554d5c22715c2f4