Fix selecting <include> fragments
We now reuse the PSI XML pull parser in layoutlib for included
layouts. This has the advantage that it picks up unsaved edits.
However, it meant that the ViewInfo objects from layoutlib would
point to valid XmlTag view cookies in the included layout rather
than null as before, which meant clicking on an <include>'ed
region in the layout would select an XML tag rather than treat
the whole included subtree as a leaf and select it.

Change-Id: Ia2f17ab5c2e672f0e8b599a050a9cffee491b34e