Fixed bug where default JDK field is empty in Project Structure.
The bug is as follows:
In a fresh installation of Android Studio, *without* an Android
SDK but with a JDK, open the Project Structure from the "Welcome
Page". Even though Android Studio was smart enough to create a
JDK, the "Android SDK" location page does not show the path of
such JDK, but instead is empty!

This CL fixes this bug.

Change-Id: I102760dbe368dfbdffbc7c4eb9cffbc1c084cbde
(cherry picked from commit 903b3a57f708f60b6ccfb2da6bdb6d282cac39d0)