Now we set the project Java lang level from the max version found.
In this CL we get the max Java language level from all the
modules in a project and set that level in the project. In the
past we used to take just the first one we found, without taking
into account that each module in a project may have a different
Java lang level.

The problem is not completely solved. For pure Java modules (e.g.
Java libraries) we don't have any information of the Java lang
level. The main issue is that Gradle does not provide a Java
model the same way the Android plug-in provides AndroidProject.

We decided not long ago that we will provide a Java-only
module from Gradle. Once that is done we'll obtain the Java
lang level from it and set it properly in the project.

Change-Id: I953058974a9c8f3aed9d4ef2c938a50e994b898a