Handle app namespaces in the layout editor
If you're editing a custom view (such as the v7 GridLayout) and use
the property sheet to edit its properties, we need to use (and declare
if necessary) the app namespace rather than assuming these should be
in the android namespace. This changeset handles that, along with some
other namespace handling cleanup.

Fixes 64536: Wrong namespace used with v7 support GridLayout

Change-Id: Iee2084591262e4c0433a400c122a0d4169bf42a2