Improved detection of unsupported plug-in version.
There is a specific combination of Gradle and plug-in versions
which is not detected by Studio and not properly marked as
error: when using an unsupported version of the plug-in with
a recent version of Gradle.

The Gradle plug-in will complain about the Gradle version,
bypassing the plug-in version check.

Change-Id: Ic11b9861106df314aef7b9a815ebb72a896c5de6