Added ability to kill Gradle daemons.
This CL:
- Asks user if he/she would like to kill all
  Gradle daemons if the user closes the project
  while a project build is running
- Warns the user about the effects of killing
  of all daemons (all other builds happening,
  either in the command line or other instances
  of the IDE will be stopped)
- Kills all Gradle daemons and continues closing
  a project if the user decided that's what
  he/she wanted.
- Adds method GradleUtil.stopAllGradleDaemons
  that can be used by our update mechanism.

There is no other way to kill Gradle daemons.
Gradle tooling API does not provide a way to kill
daemons or at least to stop running tasks (it may
be available in Gradle 1.12.)
There is also no way to kill one particular
daemon, it is a "all or nothing" thing.

Fixes
https://code.google.com/p/android/issues/detail?id=64488

Change-Id: I35647a3a39c5ec932ee39b5f9d9bb86bef236251