Fix @NotNull assertion
AndroidTargetComboBoxItem (inner class of ConfigureAndroidModuleStep)
has a static method getLabel(IAndroidTarget) which takes a target and
uses the SDK to resolve it to a human-readable string. The method is
annotated as @NotNull but it used to return SdkVersionInfo.getAndroidName
which is marked as @Nullable since it returns null if the input is unknown.

This CL adds a catch for this null return and creates a label with
"Unknown API level: <API Target>"

Fixes https://code.google.com/p/android/issues/detail?id=64416

Change-Id: I91d2a37dea2c70a710d748ffbf0737ecdb660c6c