Show error for NPW when target is not writeable
This change adds better error checking to the New Project Wizard
so that it fails quickly if it cannot write into the target location
for the project. This prevents more cryptic error messages from
developing later in the process.

This change also adds better path validation to the NPW so that
the user is not allowed to enter a non-writeable path in the
first place.

Fixes: https://code.google.com/p/android/issues/detail?id=65158
Change-Id: Ia40ca9a7c3a6ae2643d53278964030a015a1cb82