logcat: Add option to show logs from the launched app only
This CL adds an option to the launch configuration where users
can specify that logcat should automatically set a filter so that
only the logs from the launched app are visible. This option is
enabled by default. When enabled, logcat will automatically
create a filter by the package name of the app and apply that.

Fixes https://code.google.com/p/android/issues/detail?id=64901

Change-Id: Id81926bcf80e91ad2d5578834ae565cae5dc59a2