Allow dumpsys actions to work globally
The dumpsys actions in the DDMS panel are activated only if
a particular client has been selected. In that case, they would
provide information specifically for that client alone.

This CL makes the dumpsys actions to be available even if
no clients are selected. In such a case, information for the
whole system is provided.

Change-Id: I27b844e94218f2f75b114e5b7a6849c9f1c38f06