Fixed when "generate sources" is invoked.
Problem:
After each Gradle project sync we *always* invoke Gradle
to generate sources. We also sync automatically when we
successfully build a project with a stale model. The
problem is that when this automatic sync happens, it also
triggers a "generate sources", which is unnecessary because
we just built the project.

This CL fixes that by specifying that we don't want to
generate sources after a successful project build triggers
a Gradle sync.

Fixes:
https://code.google.com/p/android/issues/detail?id=65568

Change-Id: I753278a91aef2920377605ca45be4350f9de8788