Asynchronous Asset Generation
Previously, we generated images for the preview windows every
time the user changed a parameter, resulting in lots of calls
to the image generator which would block the UI thread.

This change moves the generation to a util.timer thread and delays
the execution for 200ms while waiting for any other changes that
may occur.

We also no longer require image asset names to be unique across
the entire project, now they are only required to be unique in the
module that the asset wizard is targeting.

Fixes issue 65780
Change-Id: I3ad3730562615eb846fea3f70bb70a1412e1523e