More granular resource resolution
When looking for the existence of an XML resource file we
look at all files in the entire project. This leads to false positives
for our existence and uniqueness checks that should be done on a module-level.

This change calls into the AppResourceRepository code which has already solved
this problem nicely, eliminating these false positives.

Fixes https://code.google.com/p/android/issues/detail?id=58879
Change-Id: I6c2160a32c417602d6ed9a71612327ad3867bf79