Rejig analysis so that Navigation Editor can self populate when it is first Launched
Instead of requiring that the Navigation Editor be started
from a valid navigation file, make it possible to launch the editor
from a project where it's not been used before.

This is done here by scanning the project's AndroidManifest.xml file -
to get the Activity root set - from which the transition analysis
(parsing of Java files) is initiated. This procedure is re-invoked each
time a code edit completes, so as to pick up any new Activities/transitions
that are created outside the Navigation editor (or IDE in fact).

A simple graph layout algorithm is included in this CL to give a reasonable
layout of Activities, such that related Activities are kept close to each other
and transitions between Activities are kept reasonably short.

Edits to the locations of Activities (and menus) in the Navigation Editor continue to be
stored in XML files associated with the project and continue to override the locations
that are mechanically derived when the project is reloaded.

Change-Id: Iac9aa4ec0d7b094490eca475faee82dd11b8b1f0