Hide irrelevant files in ChooseOutputLocationStep
This change gives the FileTreeModel the ability to mask files
that it knows about if they're irrelevant to the proposed changes.

e.g. when creating an image asset, we would previously get:

res/
 - drawable/
 - menu/
 - layout/
etc...

Now, if the filter is enabled, the preview shown is just of the folders
affected. i.e.

res/
 - drawable/

Fixes https://code.google.com/p/android/issues/detail?id=65562

Change-Id: I5efe71eadd6633d0a5aa4de189a50f604ec0c05c