More on code analysis: full traversal of Transition Graph and onClick() listeners
Compute 'transitive closure' of the transition graph so as to discover all of the
Activities that are, directly or indirectly, reachable from the root Activity, as
declared in AndroidManifest.xml.

Enable 'onClick' listener detection and creation in the Navigation Editor. The effect of
this is to enable users to click on a Button in the Navigation editor and drag a link to
a new Activity so as to have the appropriate 'onClick()' listener registered to launch the new
Activity. This CL also has the parsing code to do the reverse - infer the 'onClick'
transitions from the source - whether or not those sources were created by the
Navigation Editor.

Change-Id: I29150f8eb1bee56ed769fb0757c82f80d68d7268