Fixed infinite loop when setting project-level Java language version.
Problem:
When having a project that specified Java language version 1.7 *and*
the Studio has JDK 6 specified, an infinite loop occurred, and the
user was continuously asked if she wants to apply Java language
changes.

Solution:
Now Studio checks that the installed JDK is appropriate for the
project's Java language version, and if it is not it tries to
create one. If Studio cannot create an appropriate JDK, it will
prompt the user to specify the path of a JDK.

Fixes:
https://code.google.com/p/android/issues/detail?id=65145

Change-Id: I740bec7ff1fd0668bac1147db2d008f120f1c03e