Now we remove all project's libraries before importing a project.
Use case:
Import a project created with Studio 0.4.3 or earlier. That version
of Studio had a bug where it created multiple copies of the same
library. We fixed that in 0.4.4 by removing all libraries before
syncing. The problem is that we were not doing this when importing
a project.

In this CL we delete the folder .idea/libraries if it present. We
rather do this than use IDEA APIs. It is safe to do this for
project import because we delete this folder *before* creating the
actual Project object, so we don't mess up with the project's
state.

Change-Id: Ie1a4c45f90c67e84f67236c630dcf34b7bd37a79