Obtain activities for app by iterating through source roots
During the app launch process, we need access to the list of
all activities and activity aliases for the entire app. An earlier
CL obtained this information using the merged manifest file. However,
this led to occasional problems when the merged manifest file isn't
there (it is generated during build), or if the IDE has a stale copy
of it (Issue 65698).

This CL obtains the same information by looking at all the source
manifests. This is not exactly ideal since it diverges from what the
real manifest merger does. However, we have to stick with this until
we get a proper API for the manifest merger. (This task is tracked
in https://code.google.com/p/android/issues/detail?id=66388).

This CL fixes https://code.google.com/p/android/issues/detail?id=65698

Change-Id: I7b3380d12f1c2a678774db9eb91fc74562746411