Target app package if not specified
Now if no package is selected when the wizard is invoked, the object
will be placed in the application package.

Previously we would reload the template variables when selecting
templates from the list, this overwrote the calculated value of the
app package. Add the package to the final list prevents this.

Fixes https://code.google.com/p/android/issues/detail?id=65981

Change-Id: Ib6f32d66cb25966dfd0071f4865813cf58b425d7