Studio now handles unresolved dependencies in Java modules.
Problem:
For Java-library-only modules, Studio relies on Gradle's own
abstraction of an IDEA module. This abstraction does not have
a notion of "unresolved dependency". Instead, all
dependencies, resolved and unresolved, are treated almost
equally. The difference is that for the unresolved ones, the
string "unresolved -" is added at the beginning of the
dependency name. Studio right now does not know how to handle
this, and treats all dependencies as resolved ones, and never
reports the unresolved ones.

Solution:
AndroidGradleProjectResolver now parses library names for
dependencies in Java-only modules, and reports any unresolved
dependency.

Fixes
https://code.google.com/p/android/issues/detail?id=66354

Change-Id: I7818569ad274b5842ec8293f253286361d487a7f