Add 'External' Parameter type
There are some use cases where the developer must enter information
from an external source. The cannonical example of this is obtaining
an API key. This change adds an 'External' type to the parameters
which attempts to facilitate this process.

The parameter has a mandatory attribute; 'href'
e.g.
<parameter
	id="apiKey"
	name="API Key"
	type="external"
	href="https://console.developers.google.com" />

This will render as a JTextField paired with a JButton. Clicking
the JButton will attempt to open the browser and navigate to the sourceUrl.
If browsing is not permitted, a dialog will be shown to the user with
instructions.

Change-Id: I8f6659995e9fc133b30ced6a2d4d996776e03187