Adds debugKeystorePath param and keystoreSha1 method.
This allows template developers to populate parameter fields with the
SHA1 hash of the debug keystore. For example:

suggest="${keystoreSha1(debugKeystorePath)}"

Many API providers (e.g. Google, Facebook, Foursquare) require the
keystore hash to generate an API key. Providing it as part of the
template system will make it easier for Android developers to sign up
to access these APIs.

Change-Id: Ie25a60ba14f1c8d208e8f3e43e13249895551efe